requirements.txt
feedparser>=6.0.0
tmdbv3api>=1.9.0
python-dotenv>=1.0.0
schedule>=1.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0
bencodepy>=0.9.5
python-telegram-bot>=20.7
